Auto_Blue_Side - Color sensor reading tests

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/libs/RobotInit.java
@@ -49,7 +49,7 @@ public class RobotInit{
     private double motorPower;
     HardwareMap hwMap = null;
 
-    public void setMotorPower(int power){
+    public void setMotorPower(double power){
         motorPower = power;
         frontRightDrive.setPower(power);
         frontLeftDrive.setPower(power);
